+++ src/algorithms/mod.rs
@@ -7,6 +7,7 @@
 pub mod community;
 pub mod connectivity;
 pub mod constructors;
+pub mod flow;
 pub mod games;
 pub mod io;
 pub mod operators;